chore: prepare Release 48 for release testing
