+++ meteor/server/migration/X_X_X.ts
@@ -17,35 +17,4 @@ import { unprotectString } from '@sofie-automation/corelib/dist/protectedString'
 
 export const addSteps = addMigrationSteps(CURRENT_SYSTEM_VERSION, [
 	// Add some migrations!
-	{
-		id: `TranslationBundles originId`,
-		canBeRunAutomatically: true,
-		validate: () => {
-			const objects = TranslationsBundles.find({
-				blueprintOriginId: { $exists: true },
-			}).count()
-			if (objects > 0) {
-				return `object needs to be converted`
-			}
-			return false
-		},
-		migrate: () => {
-			const objects = TranslationsBundles.find({
-				blueprintOriginId: { $exists: true },
-			}).fetch()
-			for (const obj0 of objects) {
-				const obj = obj0 as TranslationsBundle & { blueprintOriginId: BlueprintId }
-				const id = generateTranslationBundleOriginId(obj.blueprintOriginId, 'blueprints')
-				TranslationsBundles.update(obj._id, {
-					$set: {
-						originId: id,
-						namespace: unprotectString(id),
-					},
-					$unset: {
-						blueprintOriginId: 1,
-					},
-				})
-			}
-		},
-	},
 ])
