+++ meteor/server/migration/migrations.ts
@@ -44,6 +44,9 @@ addSteps1_46_0()
 import { addSteps as addSteps1_47_0 } from './1_47_0'
 addSteps1_47_0()
 
+import { addSteps as addSteps1_48_0 } from './1_48_0'
+addSteps1_48_0()
+
 // Migrations for the in-development release:
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
